edit doc for chapter 7. start at 7.3 signed and unsigned characters.
What would be better way to document a book like this? just summarize
everything seems not a good thing. But if I document too abstractly, it
would be the same.

Frequent re-reading & editing is a solution for now, but I feel more
need to specify it. One thing is to write "find things when you need
it" for things too specific like type casting notation rules depend
on version of a language, what type, etc.

Don't try to write everything well at the first time.