edit doc for selection chapter. too verbose. need an edit plan:
short name for repetitive words shows too often. decide convention when
to use or not to use.

during reread, it was hard to understand the contents. considering docs
written in english, using only english could be better.

examples and forms are not illustrative enough. if contents in the book
isn't enough, consider make my own.

don't need to cover every sections and elements. rather, choose what to
document, think how to illustrate the contents well. don't just write
down the same as the original book.